+++ source/App/source/ExampleHost.FunctionApp02/Functions/SubsystemAuthenticationFunction.cs
@@ -23,6 +23,10 @@ namespace ExampleHost.FunctionApp02.Functions;
 /// <summary>
 /// This class is used for the subsystem-to-subsystem communication scenario (server side)
 /// </summary>
+/// <remarks>
+/// Similar functionality exists for Web App in the 'SubsystemAuthenticationController' class
+/// located in the 'ExampleHost.WebApi02' project.
+/// </remarks>
 public class SubsystemAuthenticationFunction
 {
     /// <summary>