+++ source/App/source/ExampleHost.WebApi.Tests/Integration/AuthenticationTests.cs
@@ -12,12 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.IdentityModel.Tokens.Jwt;
 using System.Net;
-using System.Security.Claims;
 using ExampleHost.WebApi.Tests.Fixtures;
 using FluentAssertions;
-using Microsoft.IdentityModel.Tokens;
 using Xunit;
 
 namespace ExampleHost.WebApi.Tests.Integration;